Added hosted service logic for download service and refactoring

+++ Backend/UserManager.Application/Commands/Handlers/DeleteUserCommandHandler.cs
@@ -1,15 +1,23 @@
-﻿using Application.Commands.Requests;
+﻿using UserManager.Application.Commands.Requests;
 using MediatR;
 using UserManager.Domain.Interfaces;
 
-
-namespace Application.Commands.Handlers
+namespace UserManager.Application.Commands.Handlers
 {
-    public class DeleteUserCommandHandler(IUserRepository userRepository) : IRequestHandler<DeleteUserCommand>
+    public class DeleteUserCommandHandler(IUserRepository userRepository) : IRequestHandler<DeleteUserCommand, bool>
     {
-        public async Task Handle(DeleteUserCommand request, CancellationToken cancellationToken)
+        public async Task<bool> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
         {
+            var existingUser = await userRepository.GetByIdAsync(request.Id);
+
+            if (existingUser == null)
+            {
+                return false;
+            }
+
             await userRepository.DeleteAsync(request.Id);
+
+            return true;
         }
     }
 }
